Module 6 Final Project: Update 4 - GUI (JavaFX)
This code includes GUI (JavaFX) to run the code and it displays a message that prompts the user to enter their friend's name and date. Below these buttons, two more buttons are for saving and for showing the birthday. It lets the user know how many days until the friend's birthday is, when it is 0, happy birthday to your friend! Go celebrate!